+++ app/database/migrations/2014_09_25_175640_add_image_to_entranxits_table.php
@@ -12,10 +12,14 @@ class AddImageToEntranxitsTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::table('entranxits', function(Blueprint $table)
-		{
-			$table->string('image');
-		});
+      if (App::environment() == 'testing') {
+         // No need in testing environment
+      } else {
+         Schema::table('entranxits', function(Blueprint $table)
+         {
+            $table->string('image');
+         });
+      }
 	}
 
 
@@ -26,10 +30,14 @@ class AddImageToEntranxitsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::table('entranxits', function(Blueprint $table)
-		{
-			$table->dropColumn('image');
-		});
+      if (App::environment() == 'testing') {
+         // No need in testing environment
+      } else {
+         Schema::table('entranxits', function(Blueprint $table)
+         {
+            $table->dropColumn('image');
+         });
+      }
 	}
 
 }
